packages/cli/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "priorstudio"
version = "0.4.0"
description = "CLI for the PriorStudio framework — scaffold, validate, lint, and run prior-fitted foundation model projects."
readme = "../../README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "PriorStudio contributors" }]
keywords = ["foundation-models", "pfn", "prior-fitted-networks", "causal-inference", "ml"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
  "typer>=0.12",
  "rich>=13.7",
  "pyyaml>=6.0",
  "jsonschema>=4.21",
  "requests>=2.31",
  "priorstudio-core>=0.4.0",
]

[project.optional-dependencies]
torch = ["priorstudio-core[torch]"]
wandb = ["wandb>=0.16"]
mlflow = ["mlflow>=2.10"]
modal = ["modal>=0.62"]
runpod = ["runpod>=1.6"]
hf = ["huggingface_hub>=0.20"]
studio = ["priorstudio-studio>=0.4.0"]
all = [
  "priorstudio-core[torch]",
  "wandb>=0.16",
  "mlflow>=2.10",
  "modal>=0.62",
  "runpod>=1.6",
  "huggingface_hub>=0.20",
  "priorstudio-studio>=0.4.0",
]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4"]

[project.scripts]
priorstudio = "priorstudio.cli:app"

[project.urls]
Homepage = "https://github.com/profitopsai/priorstudio"
Repository = "https://github.com/profitopsai/priorstudio"
Issues = "https://github.com/profitopsai/priorstudio/issues"

[tool.hatch.build.targets.wheel]
packages = ["priorstudio"]
artifacts = [
  "priorstudio/_bundled/**/*",
]

[tool.hatch.build.targets.wheel.force-include]
"../../schemas" = "priorstudio/_bundled/schemas"
"../../templates" = "priorstudio/_bundled/templates"
